# 8318222: RISC-V: C2 CmpU3