@Hamlin-Li 2023-10-24
> LGTM. May I ask on which platform was the JMH tested? 

It's `StarFive VisionFive 2 v1.3B`

> Also I think it's safer to perform some regression tests like tier1-3.

Thanks for reminding, I've run the hotspot compiler tests, and related jdk tests found by `grep -nr test/jdk  -we Long.compareUnsigned -we Integer.compareUnsigned`